Implemented new dataLayer handling for Activity & Heartrate.
Both now send dataItems containing their current date to the handheld.
The handheld will go and fetch this data from the FitApi and return it through the dataLayer.

Since dataLayer items are synchronized, we do not need to worry about when the phone is disconnected, as they will sync when it resumes.